ecommerse/requirements.txt
asgiref==3.2.10
Django==2.0.8
django-crispy-forms==1.8.0
django-filter==2.2.0
django-multiupload==0.5.2
Pillow==7.2.0
psycopg2==2.7.7
pytz==2020.1
sqlparse==0.3.1
